requirements.txt

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
psycopg2-binary>=2.9.9
redis>=5.0.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
httpx>=0.27.0
pytest>=8.3.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
